refactor: paging의 make()부분
