+++ src/main/java/com/sandbox/domain/articles/exception/ErrorArticleResp.java
@@ -1,0 +1,10 @@
+package com.sandbox.domain.articles.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ErrorArticleResp extends RuntimeException {
+    private String message;
+}
