+++ src/main/java/com/sandbox/domain/articles/repository/ArticleRepository.java
@@ -1,17 +1,20 @@
-package com.sandbox.domain.articles.dao;
+package com.sandbox.domain.articles.repository;
 
+import com.sandbox.domain.articles.entity.Article;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @Repository
 @RequiredArgsConstructor
 @Transactional
-public class ArticleDao {
+public class ArticleRepository {
     @PersistenceContext
     private final EntityManager em;
 
@@ -25,16 +28,4 @@ public class ArticleDao {
         }
     }
 
-    public List<Article> getOffsetList() {
-        return em.createQuery("select article from Article article", Article.class).getResultList();
-    }
-
-    public List<Article> getCursorList(int size, int cursorId) {
-        return em.createQuery("select article from Article article "
-                                + "where article.id > :cursorId "
-                                + "order by article.id", Article.class)
-                .setParameter("cursorId", cursorId)
-                .setMaxResults(size)
-                .getResultList();
-    }
 }
