+++ src/main/java/com/sandbox/domain/articles/service/ArticleService.java
@@ -2,11 +2,13 @@ package com.sandbox.domain.articles.service;
 
 import com.sandbox.domain.articles.dto.ArticleCursorResp;
 import com.sandbox.domain.articles.dto.ArticleOffsetResp;
+import com.sandbox.domain.articles.dto.ArticleReq;
+import com.sandbox.domain.articles.dto.ArticleSuccessMsgResp;
 
 import java.util.List;
 
 public interface ArticleService {
-    void makeArticleList(List<Article> articleList);
-    ArticleOffsetResp getOffsetPage(int size, int page);
-    ArticleCursorResp getCursorPage(int size, int CursorId);
+    ArticleSuccessMsgResp makeArticleList(List<ArticleReq> articleList);
+//    ArticleOffsetResp getOffsetPage(int size, int page);
+//    ArticleCursorResp getCursorPage(int size, int CursorId);
 }
